refactor: alinhar estrutura do módulo anestesia às convenções
- utils/ → services/ (lógica pura no lugar correto)
- hooks/ → components/hooks/ (hooks dentro da camada de UI)
- Atualiza todos os imports relativos
