+++ src/modules/anestesia/components/hooks/use-print.ts
